Backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.3
PyJWT==2.8.0
scikit-learn==1.3.2
mlxtend>=0.23.0
pandas==2.1.4
numpy==1.26.3
requests==2.31.0
twilio==8.10.0
python-dotenv==1.0.0
